feat(wip): Parse file into blocks.
According to the parsing strategy found at spec.commonmark.com the frist
step to parsing markdown is to break the input into blocks. Since the
only thing I care about is implementing headings and paragraphs to start
with before potentially implementing the rest of the spect, starting
with recognizing ATX headings is a good start.